Move confluent_kafka_input to connector
