+++ tests/acceptance/util.py
@@ -8,7 +8,7 @@ from multiprocessing import Lock
 from os import path, makedirs
 from os.path import join
 
-from logprep.input.confluent_kafka_input import ConfluentKafkaInputFactory
+from logprep.connector.confluent_kafka_input import ConfluentKafkaInputFactory
 from logprep.output.confluent_kafka_output import ConfluentKafkaOutputFactory
 from logprep.framework.pipeline import Pipeline, SharedCounter
 from logprep.util.helper import recursive_compare, remove_file_if_exists
