+++ tests/unit/connector/test_confluent_kafka.py
@@ -17,7 +17,7 @@ from zlib import decompress
 
 
 from logprep.connector.connector_factory_error import InvalidConfigurationError
-from logprep.input.confluent_kafka_input import ConfluentKafkaInput, ConfluentKafkaInputFactory
+from logprep.connector.confluent_kafka_input import ConfluentKafkaInput, ConfluentKafkaInputFactory
 from logprep.output.confluent_kafka_output import (
     ConfluentKafkaOutput,
     UnknownOptionError,
@@ -103,13 +103,13 @@ class TestConfluentKafkaFactory:
         kafka = ConfluentKafkaInputFactory.create_from_configuration(self.config)
         assert not all(kafka._config["ssl"][key] for key in kafka._config["ssl"])
 
-    @mock.patch("logprep.input.confluent_kafka_input.Consumer")
+    @mock.patch("logprep.connector.confluent_kafka_input.Consumer")
     def test_create_consume_calls_subscribe(self, _):
         kafka = ConfluentKafkaInputFactory.create_from_configuration(self.config)
         kafka._create_consumer()
         kafka._consumer.subscribe.assert_called()
 
-    @mock.patch("logprep.input.confluent_kafka_input.Consumer")
+    @mock.patch("logprep.connector.confluent_kafka_input.Consumer")
     def test_get_next_creates_consumer_if_consumer_is_none(self, mock_consumer):
         kafka = ConfluentKafkaInputFactory.create_from_configuration(self.config)
         assert kafka._consumer is None
@@ -842,7 +842,7 @@ class TestConfluentKafka:
         ):
             _ = kafka._set_connector_type_options(user_config, default_config)
 
-    @mock.patch("logprep.input.confluent_kafka_input.Consumer")
+    @mock.patch("logprep.connector.confluent_kafka_input.Consumer")
     def test_shut_down_calls_consumer_close(self, _):
         config = deepcopy(TestConfluentKafkaFactory.valid_configuration)
         kafka = ConfluentKafkaInputFactory.create_from_configuration(config)
@@ -851,7 +851,7 @@ class TestConfluentKafka:
         kafka.shut_down()
         kafka_consumer.close.assert_called()
 
-    @mock.patch("logprep.input.confluent_kafka_input.Consumer")
+    @mock.patch("logprep.connector.confluent_kafka_input.Consumer")
     def test_shut_down_sets_consumer_to_none(self, _):
         config = deepcopy(TestConfluentKafkaFactory.valid_configuration)
         kafka = ConfluentKafkaInputFactory.create_from_configuration(config)
