+++ tests/unit/connector/test_connector_factory.py
@@ -16,7 +16,7 @@ from logprep.connector.connector_factory import (
     ConnectorFactory,
     UnknownConnectorTypeError,
 )
-from logprep.input.confluent_kafka_input import ConfluentKafkaInput
+from logprep.connector.confluent_kafka_input import ConfluentKafkaInput
 from logprep.input.dummy_input import DummyInput
 from logprep.input.json_input import JsonInput
 from logprep.input.jsonl_input import JsonlInput
