+++ tests/unit/connector/test_elasticsearch_output_factory.py
@@ -8,7 +8,7 @@ from copy import deepcopy
 
 from pytest import fail, raises
 
-from logprep.input.confluent_kafka_input import ConfluentKafkaInputFactory
+from logprep.connector.confluent_kafka_input import ConfluentKafkaInputFactory
 from logprep.output.es_output import ElasticsearchOutputFactory
 from logprep.connector.connector_factory import InvalidConfigurationError
